Do CHANGE MASTER value conversion as late as possible
By doing the conversion inside the Pinloki instance, we will end up having
the same semantics as MariaDB where a CHANGE MASTER adds to the
configuration instead of overwriting the existing configuration. In
addition, we make the whole operation atomic.

We also remove the need to expose the MasterConfig struct and only expose
an enumeration that maps to the known values. This could be further
abstracted by only returning pairs of strings but this would essentially
be parsing outside of the parser. The same could be done for the SET
parsing but there are only a few relevant values that we are interested
in.

Moved the parser into the pinloki namespace to avoid conflicts with the
x3::parser class.