Save GTID only on transaction commit
The GTID position is now updated only when the whole transaction is
processed. This prevents the writing of partial events which would occur
when the connection to the master is lost mid-transaction.

If the connection is lost, the events of the incomplete transaction will
still be on the disk and will be overwritten when the reconnection
occurs. This theoretically can lead to corrupted replication if a
different master is chosen and it logs the same GTID but with different
content.