Add binlog purge test and fix purge problems
The purging of binlogs didn't remove them from the inventory and the file
name comparison used only the filename when it should've used the absolute
path.

The test does not currently pass as it triggers a debug assertion that is
caused by deleting a file that a slave is reading.